+++ flask_monitoringdashboard/database/__init__.py
@@ -4,6 +4,7 @@ For information about how to access the database via a session-variable, see: se
 import datetime
 import random
 import time
+import uuid
 from contextlib import contextmanager
 
 from sqlalchemy import (
@@ -54,6 +55,26 @@ class User(Base):
 
     def check_password(self, password):
         return check_password_hash(self.password_hash, password)
+
+
+class TelemetryUser(Base):
+    """Table for storing a unique identifier of an FMD user"""
+    __tablename__ = '{}TelemetryUser'.format(config.table_prefix)
+
+    id = Column(String(40), primary_key=True, default=str(uuid.uuid4()))
+    """Unique anonymous identifier to group the data received through telemetry"""
+
+    times_initialized = Column(Integer, default=1)
+    """For checking the amount of times the app was initialized"""
+
+    last_initialized = Column(DateTime, default=datetime.datetime.utcnow)
+    """Check when was the last time user accessed FMD"""
+
+    survey_filled = Column(Integer, default=1)
+    """If user filled the survey 1 - not responded 2 - declined 3 - filled"""
+
+    monitoring_consent = Column(Integer, default=1)
+    """If user agrees to share data 1 - not responded 2 - declined 3 - accepted"""
 
 
 class Endpoint(Base):