+++ flask_monitoringdashboard/frontend/webpack.config.js
@@ -1,4 +1,5 @@
 const path = require('path');
+const MiniCssExtractPlugin = require('mini-css-extract-plugin');
 
 module.exports = {
     mode: process.env.NODE_ENV === 'production' ? 'production' : 'development',
@@ -6,41 +7,35 @@ module.exports = {
     output: {
         path: path.resolve(__dirname, '../static'),
         filename: 'js/app.js',
+        publicPath: '/static/'
     },
     module: {
         rules: [
             {
                 test: /\.scss$/,
                 use: [
+                    MiniCssExtractPlugin.loader,
                     {
-                        loader: 'file-loader',
+                        loader: 'css-loader',
                         options: {
-                            name: 'css/[name].css',
+                            url: false
                         }
                     },
-                    {
-                        loader: 'extract-loader'
-                    },
-                    {
-                        loader: 'css-loader?-url'
-                    },
-                    {
-                        loader: 'sass-loader'
-                    }
+                    'sass-loader'
                 ]
             },
             {
                 test: /\.(woff(2)?|ttf|eot|svg|otf)(\?v=\d+\.\d+\.\d+)?$/,
-                use: [
-                    {
-                        loader: 'file-loader',
-                        options: {
-                            name: '[name].[ext]',
-                            outputPath: 'fonts/'
-                        }
-                    }
-                ]
+                type: 'asset/resource',
+                generator: {
+                    filename: 'fonts/[name][ext]'
+                }
             }
         ]
-    }
+    },
+    plugins: [
+        new MiniCssExtractPlugin({
+            filename: 'css/[name].css',
+        })
+    ]
 };