+++ flask_monitoringdashboard/views/endpoint.py
@@ -1,6 +1,6 @@
 from flask import jsonify, request, json
 
-from flask_monitoringdashboard import blueprint
+from flask_monitoringdashboard import blueprint, config, telemetry_config
 from flask_monitoringdashboard.controllers.endpoints import (
     get_endpoint_overview,
     get_api_performance,
@@ -14,6 +14,7 @@ from flask_monitoringdashboard.controllers.requests import (
 )
 from flask_monitoringdashboard.core.auth import secure, admin_secure
 from flask_monitoringdashboard.core.utils import get_endpoint_details
+from flask_monitoringdashboard.core.telemetry import initialize_telemetry_session, post_to_back_if_telemetry_enabled
 from flask_monitoringdashboard.database import session_scope, row2dict
 from flask_monitoringdashboard.database.endpoint import (
     get_users,
@@ -23,7 +24,7 @@ from flask_monitoringdashboard.database.endpoint import (
 )
 
 
-@blueprint.route('/api/overview')
+@blueprint.route('/api/overview', methods=['GET', 'POST'])
 @secure
 def get_overview():
     """
@@ -31,6 +32,9 @@ def get_overview():
     :return: A JSON-list with a JSON-object per endpoint
     """
     with session_scope() as session:
+        if not telemetry_config.telemetry_initialized:
+            initialize_telemetry_session(session)
+        post_to_back_if_telemetry_enabled(**{'name': 'overview'})
         return jsonify(get_endpoint_overview(session))
 
 
@@ -41,6 +45,7 @@ def users(endpoint_id):
     :param endpoint_id: integer
     :return: A JSON-list with all users of a specific endpoint (user represented by a string)
     """
+    post_to_back_if_telemetry_enabled(**{'name': f'users/{endpoint_id}'})
     with session_scope() as session:
         users_hits = get_users(session, endpoint_id)
         dicts = []
@@ -56,6 +61,7 @@ def ips(endpoint_id):
     :param endpoint_id: integer
     :return: A JSON-list with all IP-addresses of a specific endpoint (ip represented by a string)
     """
+    post_to_back_if_telemetry_enabled(**{'name': f'ip/{endpoint_id}'})
     with session_scope() as session:
         ips_hits = get_ips(session, endpoint_id)
         dicts = []
@@ -71,6 +77,7 @@ def endpoints():
     :return: A JSON-list with information about every endpoint (encoded in a JSON-object)
         For more information per endpoint, see :func: get_overview
     """
+    post_to_back_if_telemetry_enabled(**{'name': 'endpoints'})
     with session_scope() as session:
         return jsonify([row2dict(row) for row in get_endpoints(session)])
 
@@ -83,6 +90,7 @@ def endpoints_hits():
     (encoded in a JSON-object)
         For more information per endpoint, see :func: get_overview
     """
+    post_to_back_if_telemetry_enabled(**{'name': 'endpoints_hits'})
     with session_scope() as session:
         end_hits = get_endpoints_hits(session)
         dicts = []
@@ -105,6 +113,7 @@ def api_performance():
           'values': [100, 101, 102, ...]
         }
     """
+    post_to_back_if_telemetry_enabled(**{'name': 'api_performance'})
     data = json.loads(request.data)['data']
     endpoints = data['endpoints']
 
@@ -118,6 +127,7 @@ def set_rule():
     """
         The data from the form is validated and processed, such that the required rule is monitored
     """
+    post_to_back_if_telemetry_enabled(**{'name': 'set_rule'})
     endpoint_name = request.form['name']
     value = int(request.form['value'])
     with session_scope() as session:
@@ -139,6 +149,7 @@ def endpoint_info(endpoint_id):
         - total_hits: number of hits
         - url: link to this endpoint
     """
+    post_to_back_if_telemetry_enabled(**{'name': f'endpoint_info/{endpoint_id}'})
     with session_scope() as session:
         return jsonify(get_endpoint_details(session, endpoint_id))
 
@@ -146,6 +157,7 @@ def endpoint_info(endpoint_id):
 @blueprint.route('api/endpoint_status_code_distribution/<endpoint_id>')
 @secure
 def endpoint_status_code_distribution(endpoint_id):
+    post_to_back_if_telemetry_enabled(**{'name': f'endpoint_status_code_distribution/{endpoint_id}'})
     with session_scope() as session:
         return jsonify(get_status_code_distribution(session, endpoint_id))
 
@@ -153,6 +165,7 @@ def endpoint_status_code_distribution(endpoint_id):
 @blueprint.route('api/endpoint_status_code_summary/<endpoint_id>')
 @secure
 def endpoint_status_code_summary(endpoint_id):
+    post_to_back_if_telemetry_enabled(**{'name': f'endpoint_status_code_summary/{endpoint_id}'})
     with session_scope() as session:
         result = {
             'distribution': get_status_code_distribution(session, endpoint_id),
@@ -166,6 +179,7 @@ def endpoint_status_code_summary(endpoint_id):
 @blueprint.route('api/endpoint_versions/<endpoint_id>', methods=['POST'])
 @secure
 def endpoint_versions(endpoint_id):
+    post_to_back_if_telemetry_enabled(**{'name': f'endpoint_versions/{endpoint_id}'})
     with session_scope() as session:
         data = json.loads(request.data)['data']
         versions = data['versions']
@@ -175,6 +189,7 @@ def endpoint_versions(endpoint_id):
 @blueprint.route('/api/endpoint_users/<endpoint_id>', methods=['POST'])
 @secure
 def endpoint_users(endpoint_id):
+    post_to_back_if_telemetry_enabled(**{'name': f'endpoint_users/{endpoint_id}'})
     with session_scope() as session:
         data = json.loads(request.data)['data']
         users = data['users']